# public/swagger.json
{"openapi":"3.0.0","paths":{"/api/user":{"post":{"operationId":"UserController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["User"]},"get":{"operationId":"UserController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"array","items":{"$ref":"#/components/schemas/User"}}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["User"]},"delete":{"operationId":"UserController_removeAll","parameters":[],"responses":{"200":{"description":""}},"tags":["User"]}},"/api/user/{id}":{"get":{"operationId":"UserController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["User"]},"patch":{"operationId":"UserController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["User"]},"delete":{"operationId":"UserController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["User"]}},"/api/user/{id}/score":{"patch":{"operationId":"UserController_updateScore","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateScoreDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["User"]}},"/api/user/{id}/buy":{"patch":{"operationId":"UserController_buyAsset","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BuyAssetDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/User"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["User"]}},"/api/course":{"post":{"operationId":"CourseController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCourseDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Course"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Course"]},"get":{"operationId":"CourseController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"array","items":{"$ref":"#/components/schemas/Course"}}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Course"]}},"/api/course/{id}/sentence":{"post":{"operationId":"CourseController_addSentence","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSentenceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Course"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Course"]}},"/api/course/{id}/quiz":{"post":{"operationId":"CourseController_addQuiz","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateQuizDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Course"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Course"]}},"/api/course/{id}":{"get":{"operationId":"CourseController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Course"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Course"]},"patch":{"operationId":"CourseController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateCourseDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Course"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Course"]},"delete":{"operationId":"CourseController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Course"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Course"]}},"/api/asset":{"post":{"operationId":"AssetController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateAssetDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Asset"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Asset"]},"get":{"operationId":"AssetController_findAll","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"array","items":{"$ref":"#/components/schemas/Asset"}}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Asset"]}},"/api/asset/{id}":{"get":{"operationId":"AssetController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Asset"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Asset"]},"patch":{"operationId":"AssetController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateAssetDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Asset"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Asset"]},"delete":{"operationId":"AssetController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/Asset"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Asset"]}},"/api/auth/signin":{"post":{"operationId":"AuthController_signin","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignInDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/SignInDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Auth"]}},"/api/auth/signup":{"post":{"operationId":"AuthController_signUp","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignUpDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","$ref":"#/components/schemas/SignUpDto"}}}}}},"400":{"description":"","content":{"application/json":{"schema":{"properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Bad Request"}}}}}}},"tags":["Auth"]}}},"info":{"title":"Learning Corner API","description":"","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer-token":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"Authorization","description":"Enter JWT token","in":"header"}},"schemas":{"User":{"type":"object","properties":{"name":{"type":"string","example":"User name","description":"User name"},"email":{"type":"string","example":"example@gmail.com","description":"User email"},"password":{"type":"string","example":"hashed password","description":"Received the plain text password and store the hashed version"},"point":{"type":"number","example":100,"description":"User point"},"exp":{"type":"number","example":100,"description":"User experience point"},"courses":{"example":"User courses","description":"User courses","type":"array","items":{"type":"string"}},"assets":{"example":"asset _id","description":"array of user asset","type":"array","items":{"type":"string"}}},"required":["name","email","password","point","exp","courses","assets"]},"CreateUserDto":{"type":"object","properties":{"name":{"type":"string","example":"user"},"email":{"type":"string","example":"user@gmail.com"},"password":{"type":"string","example":"password"}},"required":["name","email","password"]},"UpdateScoreDto":{"type":"object","properties":{"point":{"type":"number","example":10,"description":"User score"},"exp":{"type":"number","example":10,"description":"User experience point"}},"required":["point","exp"]},"BuyAssetDto":{"type":"object","properties":{"assetId":{"type":"string","example":"66dcb6d5d58802fb1ebc93ba","description":"The example value is valid."}},"required":["assetId"]},"UpdateUserDto":{"type":"object","properties":{}},"Course":{"type":"object","properties":{"title":{"type":"string","example":"Course title","description":"Course title"},"description":{"type":"string","example":"Course description","description":"Course description"},"photo":{"type":"string","example":"http://image.com","description":"Course photol url"},"exp":{"type":"number","example":100,"description":"Course experience point"},"difficulty":{"type":"string","example":"Course difficulty","description":"Course difficulty"},"quizzes":{"example":"Course quizzes","description":"Course quizzes","type":"array","items":{"type":"string"}},"script":{"example":"Course script","description":"Course script","type":"array","items":{"type":"string"}}},"required":["title","description","photo","exp","difficulty","quizzes","script"]},"Quiz":{"type":"object","properties":{"question":{"type":"string","example":"Quiz question","description":"Quiz question"},"photo":{"type":"string","example":"http://image.com","description":"Quiz photol url"},"options":{"example":"Quiz options","description":"Quiz options","type":"array","items":{"type":"string"}},"point":{"type":"number","example":10,"description":"Quiz point"}},"required":["question","photo","options","point"]},"Sentence":{"type":"object","properties":{"speaker":{"type":"string","example":"Sentence speaker","description":"Sentence speaker"},"content":{"type":"string","example":"Sentence content","description":"Sentence content"},"audioFile":{"type":"string","example":"http://audio.com","description":"Sentence audio url"},"words":{"example":"Sentence words","description":"Sentence words","type":"array","items":{"type":"string"}}},"required":["speaker","content","audioFile","words"]},"Option":{"type":"object","properties":{"content":{"type":"string","example":"Option content","description":"Option content"},"isCorrect":{"type":"boolean","example":"Option is correct","description":"Option is correct"},"photo":{"type":"string","example":"http://image.com","description":"Option photol url"}},"required":["content","isCorrect","photo"]},"CreateCourseDto":{"type":"object","properties":{"title":{"type":"string","example":"Course title","description":"Course title"},"description":{"type":"string","example":"Course description","description":"Course description"},"photo":{"type":"string","example":"http://image.com","description":"Course photol url"},"exp":{"type":"number","example":100,"description":"Course experience point"},"difficulty":{"type":"string","example":"Beginner","description":"Course difficulty"}},"required":["title","description","photo","exp","difficulty"]},"CreateSentenceDto":{"type":"object","properties":{"speaker":{"type":"string","example":"Sentence speaker","description":"Sentence speaker"},"content":{"type":"string","example":"Sentence content","description":"Sentence content"},"audioFile":{"type":"string","example":"http://audio.com","description":"Sentence audio url"}},"required":["speaker","content","audioFile"]},"CreateQuizDto":{"type":"object","properties":{"question":{"type":"string","example":"Quiz question","description":"Quiz question"},"photo":{"type":"string","example":"http://image.com","description":"Quiz photol url"},"options":{"example":[{"content":"Option content 1","isCorrect":true,"photo":"http://image.com"},{"content":"Option content 2","isCorrect":false,"photo":"http://image.com"}],"description":"Quiz options","type":"array","items":{"type":"string"}},"point":{"type":"number","example":10,"description":"Quiz point"}},"required":["question","photo","options","point"]},"UpdateCourseDto":{"type":"object","properties":{"title":{"type":"string","example":"Course title","description":"Course title"},"description":{"type":"string","example":"Course description","description":"Course description"},"photo":{"type":"string","example":"http://image.com","description":"Course photol url"},"exp":{"type":"number","example":100,"description":"Course experience point"},"difficulty":{"type":"string","example":"Beginner","description":"Course difficulty"}}},"Asset":{"type":"object","properties":{"name":{"type":"string","example":"Asset name","description":"Asset name"},"photo":{"type":"string","example":"http://image.com","description":"Asset photol url"},"point":{"type":"number","example":10,"description":"Asset point"}},"required":["name","photo","point"]},"CreateAssetDto":{"type":"object","properties":{"name":{"type":"string","example":"Asset name","description":"Asset name"},"photo":{"type":"string","example":"http://image.com","description":"Asset photol url"},"point":{"type":"number","example":10,"description":"Asset point"}},"required":["name","photo","point"]},"UpdateAssetDto":{"type":"object","properties":{"name":{"type":"string","example":"Asset name","description":"Asset name"},"photo":{"type":"string","example":"http://image.com","description":"Asset photol url"},"point":{"type":"number","example":10,"description":"Asset point"}}},"SignUpDto":{"type":"object","properties":{"name":{"type":"string","example":"John","description":"User Name"},"email":{"type":"string","example":"John@gmail.com","description":"Last Name"},"password":{"type":"string","example":"password","description":"Password"}},"required":["name","email","password"]},"SignInDto":{"type":"object","properties":{"email":{"type":"string","example":"John@gmail.com","description":"Last Name"},"password":{"type":"string","example":"password","description":"Password"}},"required":["email","password"]}}}}